Extract pure-JVM passes-barcode-core (wpass-7xo.2)
Move the trust-claim-bearing ZXing symbol decode out of the Android
passes-barcode module into a new pure-JVM module so the isolated
still-image path and the future in-process live-camera path share one
decode implementation. Mirrors the passes-pdf / passes-pdf-core split.

- New passes-barcode-core: decodeLuminance(LuminanceSource) + the
  symbology allowlist (ROSTER_BY_ZXING_FORMAT, DECODE_HINTS). Depends on
  passes-core; adds com.google.zxing:core. No Android deps.
- passes-barcode keeps only the Bitmap -> RGBLuminanceSource adapter and
  delegates to the core. New edge: passes-barcode -> passes-barcode-core
  -> passes-core.
- Move the two pure-JVM decode tests (ZxingBarcodeSymbolDecoderTest,
  HostilePayloadFidelityTest) into the core module.
- CLAUDE.md: nine -> ten modules, new dependency edges.
