+++ settings.gradle.kts
@@ -33,6 +33,7 @@ include(":passes-core")
 include(":passes-pdf-core")
 include(":passes-isolation")
 include(":passes-pdf")
+include(":passes-barcode-core")
 include(":passes-barcode")
 include(":passes-storage")
 include(":passes-ui-core")
